Updated Vial keymap with ortho & staggered layouts
