+++ keyboards/walletburner/cajal/keymaps/vial/rules.mk
@@ -3,4 +3,3 @@ VIAL_ENABLE = yes
 LTO_ENABLE = yes
 
 VIAL_ENCODERS_ENABLE = yes
-VIAL_INSECURE = yes
