+++ keyboards/walletburner/cajal/keymaps/vial/vial.json
@@ -1,0 +1,30 @@
+{
+  "name": "Cajal",
+  "vendorId": "0x5742",
+  "productId": "0x6361",
+  "lighting": "qmk_backlight_rgblight",
+  "matrix": {
+    "rows": 4,
+    "cols": 14
+  },
+  "layouts": {
+    "labels": [
+      [
+	"Layout",
+        "Staggered",
+        "Ortho"
+      ]
+    ],
+    "keymap": [
+      ["0,0","0,1","0,2","0,3","0,4","0,5","0,6","0,7","0,8","0,9","0,10","0,11","0,12",{"x":0.25},"0,13"],
+      [{"w":1.25},"1,0\n\n\n0,0","1,1\n\n\n0,0","1,2\n\n\n0,0","1,3\n\n\n0,0","1,4\n\n\n0,0","1,5\n\n\n0,0","1,6\n\n\n0,0","1,7\n\n\n0,0","1,8\n\n\n0,0","1,9\n\n\n0,0","1,10\n\n\n0,0",{"w":1.75},"1,11\n\n\n0,0",{"x":0.25},"0,0\n\n\n\n\n\n\n\n\ne","0,1\n\n\n\n\n\n\n\n\ne"],
+      [{"w":1.75},"2,0\n\n\n0,0","2,1\n\n\n0,0","2,2\n\n\n0,0","2,3\n\n\n0,0","2,4\n\n\n0,0","2,5\n\n\n0,0","2,6\n\n\n0,0","2,7\n\n\n0,0","2,8\n\n\n0,0","2,9\n\n\n0,0",{"w":1.25},"2,10\n\n\n0,0"],
+      [{"y":-0.75,"x":12.5},"2,12"],
+      [{"y":-0.25,"w":1.25},"3,0\n\n\n0,0","3,1\n\n\n0,0",{"w":1.25},"3,2\n\n\n0,0",{"w":2.25},"3,3\n\n\n0,0",{"w":2.75},"3,6\n\n\n0,0",{"w":1.25},"3,8\n\n\n0,0",{"w":1.25},"3,9\n\n\n0,0"],
+      [{"y":-0.75,"x":11.5},"3,11","3,12","3,13"],
+      [{"y":0.5},"1,0\n\n\n0,1","1,1\n\n\n0,1","1,2\n\n\n0,1","1,3\n\n\n0,1","1,4\n\n\n0,1","1,5\n\n\n0,1","1,6\n\n\n0,1","1,7\n\n\n0,1","1,8\n\n\n0,1","1,9\n\n\n0,1","1,10\n\n\n0,1","1,11\n\n\n0,1","1,12\n\n\n0,1"],
+      ["2,0\n\n\n0,1","2,1\n\n\n0,1","2,2\n\n\n0,1","2,3\n\n\n0,1","2,4\n\n\n0,1","2,5\n\n\n0,1","2,6\n\n\n0,1","2,7\n\n\n0,1","2,8\n\n\n0,1","2,9\n\n\n0,1","2,10\n\n\n0,1","2,11\n\n\n0,1"],
+      ["3,0\n\n\n0,1","3,1\n\n\n0,1","3,2\n\n\n0,1","3,3\n\n\n0,1","3,4\n\n\n0,1","3,5\n\n\n0,1","3,6\n\n\n0,1","3,7\n\n\n0,1","3,8\n\n\n0,1","3,9\n\n\n0,1","3,10\n\n\n0,1"]
+    ]
+  }
+}
