--- keyboards/walletburner/cajal/keymaps/vial_ortho/vial.json
@@ -1,30 +1,0 @@
-{
-  "name": "Cajal Ortho",
-  "vendorId": "0x5742",
-  "productId": "0x6361",
-  "lighting": "qmk_backlight_rgblight",
-  "matrix": {
-    "rows": 4,
-    "cols": 14
-  },
-  "layouts": {
-    "labels": [
-      [
-        "Bottom Row",
-        "1x1",
-        "2x2",
-        "7u"
-      ]
-    ],
-    "keymap": [
-      ["0,0","0,1","0,2","0,3","0,4","0,5","0,6","0,7","0,8","0,9","0,10","0,11","0,12",{"x":0.25},"0,13"],
-      ["1,0","1,1","1,2","1,3","1,4","1,5","1,6","1,7","1,8","1,9","1,10","1,11","1,12",{"x":0.25},"0,0\n\n\n\n\n\n\n\n\ne","0,1\n\n\n\n\n\n\n\n\ne"],
-      ["2,0","2,1","2,2","2,3","2,4","2,5","2,6","2,7","2,8","2,9","2,10","2,11"],
-      [{"y":-0.75,"x":12.5},"2,12"],
-      [{"y":-0.25},"3,0\n\n\n0,0","3,1\n\n\n0,0","3,2\n\n\n0,0","3,3\n\n\n0,0","3,4\n\n\n0,0","3,5\n\n\n0,0","3,6\n\n\n0,0","3,7\n\n\n0,0","3,8\n\n\n0,0","3,9\n\n\n0,0","3,10\n\n\n0,0"],
-      [{"y":-0.75,"x":11.5},"3,11","3,12","3,13"],
-      [{"y":0.25},"3,0\n\n\n0,1","3,1\n\n\n0,1","3,2\n\n\n0,1","3,3\n\n\n0,1",{"w":2},"3,4\n\n\n0,1",{"w":2},"3,6\n\n\n0,1","3,8\n\n\n0,1","3,9\n\n\n0,1","3,10\n\n\n0,1"],
-      ["3,0\n\n\n0,2","3,1\n\n\n0,2",{"w":7},"3,2\n\n\n0,2","3,9\n\n\n0,2","3,10\n\n\n0,2"]
-    ]
-  }
-}
