--- keyboards/walletburner/cajal/keymaps/vial_stag/rules.mk
@@ -1,6 +1,0 @@
-VIA_ENABLE = yes
-VIAL_ENABLE = yes
-LTO_ENABLE = yes
-
-VIAL_ENCODERS_ENABLE = yes
-VIAL_INSECURE = yes
