--- keyboards/walletburner/cajal/keymaps/vial_stag/vial.json
@@ -1,20 +1,0 @@
-{
-  "name": "Cajal Staggered",
-  "vendorId": "0x5742",
-  "productId": "0x6361",
-  "lighting": "qmk_backlight_rgblight",
-  "matrix": {
-    "rows": 4,
-    "cols": 14
-  },
-  "layouts": {
-    "keymap": [
-      ["0,0","0,1","0,2","0,3","0,4","0,5","0,6","0,7","0,8","0,9","0,10","0,11","0,12",{"x":0.25},"0,13"],
-      [{"w":1.25},"1,0","1,1","1,2","1,3","1,4","1,5","1,6","1,7","1,8","1,9","1,10",{"w":1.75},"1,11",{"x":0.25},"0,0\n\n\n\n\n\n\n\n\ne","0,1\n\n\n\n\n\n\n\n\ne"],
-      [{"w":1.75},"2,0","2,1","2,2","2,3","2,4","2,5","2,6","2,7","2,8","2,9",{"w":1.25},"2,10"],
-      [{"y":-0.5,"x":12.5},"2,12"],
-      [{"y":-0.5,"w":1.25},"3,0","3,1",{"w":1.25},"3,2",{"w":2.25},"3,3",{"w":2.75},"3,6",{"w":1.25},"3,8",{"w":1.25},"3,9"],
-      [{"y":-0.5,"x":11.5},"3,11","3,12","3,13"]
-    ]
-  }
-}
